AiScraperV1/structuredAI/requirements.txt
python-dotenv
playwright
langchain-core
langchain-google-genai
pydantic